Refactor code & Finish LibraryScreen
